# auth_service/docs/swagger.yaml
basePath: /api/auth
definitions:
  controller.RegisterRequest:
    properties:
      email:
        description: 邮箱（用于后续验证/找回密码）
        example: test@xxx.com
        type: string
      password:
        description: 密码（建议包含大小写字母+数字+符号）
        example: 123456aB!
        type: string
      username:
        description: 用户名（唯一）
        example: test_user
        type: string
    type: object
  controller.RegisterResponse:
    properties:
      code:
        description: 状态码
        example: 200
        type: integer
      message:
        description: 提示信息
        example: 注册成功
        type: string
    type: object
  model.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 路由分发，统计鉴权
  termsOfService: http://hostip/
  title: bili_auth_service
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户名密码验证通过后，生成JWT令牌并放回用户基础信息
      parameters:
      - description: 登录请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功(含token)
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.RegisterResponse'
              type: object
        "400":
          description: 请求参数错误或密码错误
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.APIResponse'
      summary: 用户登录
      tags:
      - auth
  /auth/logout:
    post:
      description: 客户端登出
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.RegisterResponse'
              type: object
      summary: 用户登出
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 接收用户名、密码、邮箱，完成用户注册
      parameters:
      - description: 注册请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.RegisterResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.APIResponse'
      summary: 用户注册
      tags:
      - auth
swagger: "2.0"